fixup: move FindAuditLine() to auditlog_helpers
