+++ ydb/core/tx/schemeshard/ut_helpers/auditlog_helpers.cpp
@@ -4,6 +4,7 @@
 
 #include <library/cpp/logger/backend.h>
 #include <library/cpp/logger/record.h>
+#include <library/cpp/testing/unittest/registar.h>
 
 #include <ydb/core/protos/config.pb.h>
 
@@ -35,6 +36,18 @@ NAudit::TAuditLogBackends CreateTestAuditLogBackends(std::vector<std::string>& l
     NAudit::TAuditLogBackends logBackends;
     logBackends[NKikimrConfig::TAuditConfig::TXT].emplace_back(new TMemoryLogBackend(lineBuffer));
     return logBackends;
+}
+
+std::string FindAuditLine(const std::vector<std::string>& auditLines, const std::string& substr) {
+    Cerr << "AUDIT LOG buffer(" << auditLines.size() << "):" << Endl;
+    for (auto i : auditLines) {
+        Cerr << "    " << i << Endl;
+    }
+    auto found = std::find_if(auditLines.begin(), auditLines.end(), [&](auto i) { return i.contains(substr); });
+    UNIT_ASSERT_C(found != auditLines.end(), "No audit record with substring: '" + substr + "'");
+    auto line = *found;
+    Cerr << "AUDIT LOG checked line:" << Endl << "    " << line << Endl;
+    return line;
 }
 
 }  // namespace NSchemeShardUT_Private
