+++ ydb/core/tx/schemeshard/ut_restore/ut_restore.cpp
@@ -50,18 +50,6 @@ namespace {
         Aws::ShutdownAPI(Options);
     }
 
-    std::string FindAuditLine(const std::vector<std::string>& auditLines, const std::string& substr) {
-        Cerr << "AUDIT LOG buffer(" << auditLines.size() << "):" << Endl;
-        for (auto i : auditLines) {
-            Cerr << "    " << i << Endl;
-        }
-        auto found = std::find_if(auditLines.begin(), auditLines.end(), [&](auto i) { return i.contains(substr); });
-        UNIT_ASSERT_C(found != auditLines.end(), "No audit record with substring: '" + substr + "'");
-        auto line = *found;
-        Cerr << "AUDIT LOG checked line:" << Endl << "    " << line << Endl;
-        return line;
-    }
-
     const TString EmptyYsonStr = R"([[[[];%false]]])";
 
     TString GenerateScheme(const NKikimrSchemeOp::TPathDescription& pathDesc) {
